groupproj/requirements.txt
django
streamlit
djangorestframework
django-cors-headers